Test Files
These are the files for testing the script. 
Extraction.py will look for the information in the Gaussian file and extract the info to later report it in a xlsx file. This is just a variation of an old code done long ago to extract information about binding energies from a Gaussian file. This script is faster and needs less memory since I don't need to read the whole file in mem I just read line by line.  
TODO:
I need to add a routine to make a gjf file for the next calculation and test it with a large set of more files. I also need to dynamically set the variables to search depending on the type of calculation. I also need to make a function to auto run depending on the shell. I also need to make it more user-friendly.